requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
blis==0.7.4
cachetools==4.2.0
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
de-core-news-md @ https://github.com/explosion/spacy-models/releases/download/de_core_news_md-2.3.0/de_core_news_md-2.3.0.tar.gz
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-2.3.0/de_core_news_sm-2.3.0.tar.gz
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
flake8==3.8.4
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
liwc==0.5.0
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.5
notebook==6.1.5
numpy==1.19.4
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plac==1.1.3
preshed==3.0.5
prettytable==2.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pydeck==0.5.0
pyflakes==2.2.0
Pygments==2.7.3
pylint-venv==2.1.1
pyparsing==2.4.7
Pyphen==0.10.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.4
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.1
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smmap==3.0.4
spacy==2.3.5
srsly==1.0.5
streamlit==0.73.0
tabulate==0.8.7
terminado==0.9.1
testpath==0.4.4
thinc==7.4.5
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.2
validators==0.18.2
wasabi==0.8.0
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==2.0.1
jupyterlab==3.0.7